myapp/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
sqlalchemy==2.0.28
pydantic==2.6.4